nvidia-gpu: Fix thresholds for GPU_TEMP_1
Fixes thresholds for GPU_TEMP_1 to be upper critical, warning,
shutdown. Rather than lower critical, et al.

+++ src/nvidia-gpu/NvidiaGpuDevice.hpp
@@ -78,7 +78,7 @@ class GpuDevice : public std::enable_shared_from_this<GpuDevice>
 
     std::array<uint8_t, sizeof(gpu::ReadThermalParametersRequest)>
         thermalParamReqMsg{};
-    std::array<uint8_t, 3> thresholds{};
+    std::array<int32_t, 3> thresholds{};
     size_t current_threshold_index{};
 
     SensorConfigs configs;
